requirements.txt
appdirs==1.4.3
Django==1.11
djangorestframework==3.6.2
gunicorn==19.7.1
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
requests